+++ examples/scripts/DFN.py
@@ -8,13 +8,14 @@ import numpy as np
 pybamm.set_logging_level("INFO")
 
 # load model
-model = pybamm.lithium_ion.DFN()
+model = pybamm.lithium_ion.DFN({"operating mode": "voltage"})
 
 # create geometry
 geometry = model.default_geometry
 
 # load parameter values and process model and geometry
 param = model.default_parameter_values
+param["Voltage function [V]"] = 4.1
 param.process_model(model)
 param.process_geometry(geometry)