+++ examples/scripts/compare_SPM_diffusion_models.py
@@ -23,7 +23,7 @@ models = [
 
 # load parameter values and process models and geometry
 param = models[0].default_parameter_values
-param.update({"Typical current [A]": 1})
+param.update({"Current function [A]": 1})
 for model in models:
     param.process_model(model)
 
@@ -42,7 +42,7 @@ for model in models:
 
 # solve model
 solutions = [None] * len(models)
-t_eval = np.linspace(0, 0.17, 100)
+t_eval = np.linspace(0, 0.25, 100)
 for i, model in enumerate(models):
     solutions[i] = model.default_solver.solve(model, t_eval)