+++ pybamm/models/submodels/current_collector/quite_conductive_potential_pair.py
@@ -47,7 +47,7 @@ class BaseQuiteConductivePotentialPair(BasePotentialPair):
     def set_algebraic(self, variables):
 
         param = self.param
-        applied_current = param.current_with_time
+        applied_current = variables["Total current density"]
         cc_area = self._get_effective_current_collector_area()
         z = pybamm.standard_spatial_vars.z