+++ pybamm/models/submodels/external_circuit/__init__.py
@@ -1,0 +1,11 @@
+from .base_external_circuit import BaseModel, LeadingOrderBaseModel
+from .current_control_external_circuit import CurrentControl, LeadingOrderCurrentControl
+from .function_control_external_circuit import (
+    FunctionControl,
+    VoltageFunctionControl,
+    PowerFunctionControl,
+    LeadingOrderFunctionControl,
+    LeadingOrderVoltageFunctionControl,
+    LeadingOrderPowerFunctionControl,
+)
+