+++ tests/integration/test_models/test_full_battery_models/test_lead_acid/test_compare_outputs.py
@@ -21,7 +21,7 @@ class TestCompareOutputs(unittest.TestCase):
 
         # load parameter values (same for all models)
         param = models[0].default_parameter_values
-        param.update({"Typical current [A]": 1})
+        param.update({"Current function [A]": 1})
         for model in models:
             param.process_model(model)
 
@@ -67,7 +67,7 @@ class TestCompareOutputs(unittest.TestCase):
         for models in model_combos:
             # load parameter values (same for all models)
             param = models[0].default_parameter_values
-            param.update({"Typical current [A]": 1})
+            param.update({"Current function [A]": 1})
             for model in models:
                 param.process_model(model)