+++ tests/unit/test_parameters/test_update_parameters.py
@@ -52,9 +52,9 @@ class TestUpdateParameters(unittest.TestCase):
         parameter_values_update = pybamm.ParameterValues(
             chemistry=pybamm.parameter_sets.Marquis2019
         )
-        parameter_values_update.update({"Typical current [A]": 2})
+        parameter_values_update.update({"Current function [A]": 1})
         modeltest2.test_update_parameters(parameter_values_update)
-        self.assertEqual(model2.variables["Current [A]"].evaluate(), 2)
+        self.assertEqual(model2.variables["Current [A]"].evaluate(), 1)
         modeltest2.test_solving(t_eval=t_eval)
         Y2 = modeltest2.solution.y
 
@@ -68,12 +68,14 @@ class TestUpdateParameters(unittest.TestCase):
         parameter_values_update = pybamm.ParameterValues(
             chemistry=pybamm.parameter_sets.Marquis2019
         )
-        parameter_values_update.update({"Current function": "[zero]"})
+        parameter_values_update.update({"Current function [A]": 0})
         modeltest3.test_update_parameters(parameter_values_update)
         modeltest3.test_solving(t_eval=t_eval)
         Y3 = modeltest3.solution.y
 
-        self.assertIsInstance(model3.variables["Current [A]"], pybamm.Scalar)
+        self.assertIsInstance(model3.variables["Current [A]"], pybamm.Multiplication)
+        self.assertIsInstance(model3.variables["Current [A]"].left, pybamm.Scalar)
+        self.assertIsInstance(model3.variables["Current [A]"].right, pybamm.Scalar)
         self.assertEqual(model3.variables["Current [A]"].evaluate(), 0.0)
 
         # results should be different
@@ -94,8 +96,14 @@ class TestUpdateParameters(unittest.TestCase):
         # test on simple lead-acid model
         model1 = pybamm.lead_acid.LOQS()
         modeltest1 = tests.StandardModelTest(model1)
+        parameter_values = pybamm.ParameterValues(
+            chemistry=pybamm.parameter_sets.Sulzer2019
+        )
+        parameter_values.update({"C-rate": 0.05})
         t_eval = np.linspace(0, 0.5)
-        modeltest1.test_all(t_eval=t_eval, skip_output_tests=True)
+        modeltest1.test_all(
+            param=parameter_values, t_eval=t_eval, skip_output_tests=True
+        )
 
         T1, Y1 = modeltest1.solution.t, modeltest1.solution.y
 
@@ -105,6 +113,7 @@ class TestUpdateParameters(unittest.TestCase):
         )
         parameter_values_update.update(
             {
+                "C-rate": 0.05,
                 "Negative electrode thickness [m]": 0.0002,
                 "Separator thickness [m]": 0.0003,
                 "Positive electrode thickness [m]": 0.0004,